+++ src/common/Types.ts
@@ -31,7 +31,7 @@ export interface FlexboxStyle {
     alignItems?: 'flex-start' | 'flex-end' | 'center' | 'stretch';
     alignSelf?: 'auto' | 'flex-start' | 'flex-end' | 'center' | 'stretch';
     alignContent?: 'auto' | 'flex-start' | 'flex-end' | 'center' | 'stretch';
-    
+
     borderWidth?: number;
     borderTopWidth?: number;
     borderRightWidth?: number;
@@ -46,6 +46,9 @@ export interface FlexboxStyle {
     bottom?: number;
     left?: number;
 
+    flexGrow?: number;
+    flexShrink?: number;
+    flexBasis?: number;
     flex?: number;
 
     flexWrap?: 'wrap' | 'nowrap';
@@ -76,14 +79,25 @@ export interface FlexboxStyle {
     position?: 'absolute' | 'relative';
 }
 
-export interface AnimatedFlexboxStyle {
-    height?: RX.AnimatedValue;
-    width?: RX.AnimatedValue;
+export abstract class AnimatedValue implements RX.IAnimatedValue {
+    constructor(val: number) {
+        // No-op
+    }
+    abstract setValue(value: number): void;
+    abstract addListener(callback: any): number;
+    abstract removeListener(id: string): void;
+    abstract removeAllListeners(): void;
+    abstract interpolate(config: any): AnimatedValue;
+}
 
-    top?: RX.AnimatedValue;
-    right?: RX.AnimatedValue;
-    bottom?: RX.AnimatedValue;
-    left?: RX.AnimatedValue;
+export interface AnimatedFlexboxStyle {
+    height?: AnimatedValue;
+    width?: AnimatedValue;
+
+    top?: AnimatedValue;
+    right?: AnimatedValue;
+    bottom?: AnimatedValue;
+    left?: AnimatedValue;
 }
 
 // ------------------------------------------------------------
@@ -107,16 +121,16 @@ export interface TransformStyle {
 
 export interface AnimatedTransformStyle {
     transform?: [{
-        perspective?: RX.AnimatedValue;
-        rotate?: RX.AnimatedValue;
-        rotateX?: RX.AnimatedValue;
-        rotateY?: RX.AnimatedValue;
-        rotateZ?: RX.AnimatedValue;
-        scale?: RX.AnimatedValue;
-        scaleX?: RX.AnimatedValue;
-        scaleY?: RX.AnimatedValue;
-        translateX?: RX.AnimatedValue;
-        translateY?: RX.AnimatedValue;
+        perspective?: AnimatedValue;
+        rotate?: AnimatedValue;
+        rotateX?: AnimatedValue;
+        rotateY?: AnimatedValue;
+        rotateZ?: AnimatedValue;
+        scale?: AnimatedValue;
+        scaleX?: AnimatedValue;
+        scaleY?: AnimatedValue;
+        translateX?: AnimatedValue;
+        translateY?: AnimatedValue;
     }];
 }
 
@@ -140,9 +154,9 @@ export interface ViewAndImageCommonStyle extends FlexboxStyle, TransformStyle {
 }
 
 export interface AnimatedViewAndImageCommonStyle extends AnimatedFlexboxStyle, AnimatedTransformStyle {
-    borderRadius?: RX.AnimatedValue;
-    backgroundColor?: RX.AnimatedValue;
-    opacity?: RX.AnimatedValue;
+    borderRadius?: AnimatedValue;
+    backgroundColor?: AnimatedValue;
+    opacity?: AnimatedValue;
 }
 
 // ------------------------------------------------------------
@@ -150,7 +164,7 @@ export interface AnimatedViewAndImageCommonStyle extends AnimatedFlexboxStyle, A
 // ------------------------------------------------------------
 
 export interface ViewStyle extends ViewAndImageCommonStyle {
-    borderStyle?: 'none' | 'solid' | 'dotted' | 'dashed';
+    borderStyle?: 'solid' | 'dotted' | 'dashed' | 'none';
     wordBreak?: 'break-all' | 'break-word'; // Web only
     appRegion?: 'drag' | 'no-drag'; // Web only
     cursor?: 'pointer' | 'default'; // Web only
@@ -241,8 +255,8 @@ export interface TextStyle extends ViewStyle {
 export type TextStyleRuleSet = StyleRuleSet<TextStyle>;
 
 export interface AnimatedTextStyle extends AnimatedViewAndImageCommonStyle {
-    color?: RX.AnimatedValue;
-    fontSize?: RX.AnimatedValue;
+    color?: AnimatedValue;
+    fontSize?: AnimatedValue;
 }
 
 export type AnimatedTextStyleRuleSet = StyleRuleSet<AnimatedTextStyle>;
@@ -257,8 +271,8 @@ export interface TextInputStyle extends TextStyle {
 export type TextInputStyleRuleSet = StyleRuleSet<TextInputStyle>;
 
 export interface AnimatedTextInputStyle extends AnimatedViewAndImageCommonStyle {
-    color?: RX.AnimatedValue;
-    fontSize?: RX.AnimatedValue;
+    color?: AnimatedValue;
+    fontSize?: AnimatedValue;
 }
 
 export type AnimatedTextInputStyleRuleSet = StyleRuleSet<AnimatedTextInputStyle>;
@@ -412,6 +426,7 @@ export interface ButtonProps extends CommonStyledProps<ButtonStyleRuleSet>, Comm
     delayLongPress?: number;
     cursor?: string;
 
+    onAccessibilityTapIOS?: Function; // iOS-only prop, call when a button is double tapped in accessibility mode
     onContextMenu?: (e: SyntheticEvent) => void;
     onPress?: (e: SyntheticEvent) => void;
     onPressIn?: (e: SyntheticEvent) => void;
@@ -525,6 +540,7 @@ export interface ViewPropsShared extends CommonProps, CommonAccessibilityProps {
     children?: ReactNode;
 
     restrictFocusWithin?: boolean; // Web-only, during the keyboard navigation, the focus will not go outside this view
+    limitFocusWithin?: boolean; // Web-only, make the view and all focusable subelements not focusable when isFocusLimited state is true
 
     importantForLayout?: boolean; // Web-only, additional invisible DOM elements will be added to track the size changes faster
 
@@ -535,6 +551,7 @@ export interface ViewPropsShared extends CommonProps, CommonAccessibilityProps {
     animateChildLeave?: boolean;
     animateChildMove?: boolean;
 
+    onAccessibilityTapIOS?: Function;
     onLayout?: (e: ViewOnLayoutEvent) => void;
     onMouseEnter?: (e: MouseEvent) => void;
     onMouseLeave?: (e: MouseEvent) => void;
@@ -971,7 +988,7 @@ export interface NavigatorProps extends CommonProps {
     renderScene: (route: NavigatorRoute) => JSX.Element;
     navigateBackCompleted?: () => void;
     // NOTE: Arguments are only passed to transitionStarted by the experimental navigator
-    transitionStarted?: (progress?: RX.AnimatedValue,
+    transitionStarted?: (progress?: RX.IAnimatedValue,
         toRouteId?: string, fromRouteId?: string,
         toIndex?: number, fromIndex?: number) => void;
     transitionCompleted?: () => void;
@@ -1040,7 +1057,7 @@ export module Animated {
         outputRange: (number | string)[];
     }
 
-    export type TimingFunction = (value: RX.AnimatedValue, config: TimingAnimationConfig) => CompositeAnimation;
+    export type TimingFunction = (value: RX.IAnimatedValue, config: TimingAnimationConfig) => CompositeAnimation;
     export var timing: TimingFunction;
 
     export type SequenceFunction = (animations: Array<CompositeAnimation>) => CompositeAnimation;
@@ -1162,7 +1179,8 @@ export interface LinkingErrorInfo {
 export enum AppActivationState {
     Active = 1,
     Background = 2,
-    Inactive = 3
+    Inactive = 3,
+    Extension = 4
 }
 
 // UserInterface