+++ src/native-common/View.tsx
@@ -100,6 +100,14 @@ export class View extends ViewBase<Types.ViewProps, {}> {
     focus() {
         AccessibilityUtil.setAccessibilityFocus(this);
     }
+
+    setFocusRestricted(restricted: boolean) {
+        // Nothing to do.
+    }
+
+    setFocusLimited(limited: boolean) {
+        // Nothing to do.
+    }
 }
 
 export default View;