+++ src/web/utils/FocusManager.ts
@@ -16,16 +16,23 @@ let _lastComponentId: number = 0;
 interface StoredFocusableComponent {
     component: React.Component<any, any>;
     onFocus: EventListener;
-    restricted?: boolean;
+    restricted: boolean;
+    limitedCount: number;
     origTabIndex?: number;
     removed?: boolean;
+    callbacks?: FocusableComponentStateCallback[];
 }
 
+export type FocusableComponentStateCallback = (restrictedOrLimited: boolean) => void;
+
 export class FocusManager {
+    private _parent: FocusManager;
+
     private static _rootFocusManager: FocusManager;
 
     private static _restrictionStack: FocusManager[] = [];
     private static _currentRestrictionOwner: FocusManager;
+    private _isFocusLimited: boolean;
 
     private static _currentFocusedComponent: StoredFocusableComponent;
     private _prevFocusedComponent: StoredFocusableComponent;
@@ -34,11 +41,14 @@ export class FocusManager {
 
     private _myFocusableComponentIds: { [id: string]: boolean } = {};
 
-    private _isRoot: boolean;
-
-    setAsRootFocusManager() {
-        FocusManager._rootFocusManager = this;
-        this._isRoot = true;
+    constructor(parent: FocusManager) {
+        if (parent) {
+            this._parent = parent;
+        } else if (FocusManager._rootFocusManager) {
+            console.error('FocusManager: root is already set');
+        } else {
+            FocusManager._rootFocusManager = this;
+        }
     }
 
     // Whenever the focusable element is mounted, we let the application
@@ -49,50 +59,69 @@ export class FocusManager {
             return;
         }
 
-        let componentId: string = 'fc-' + ++_lastComponentId;
+        const componentId: string = 'fc-' + ++_lastComponentId;
 
         let storedComponent: StoredFocusableComponent = {
             component: component,
+            restricted: false,
+            limitedCount: 0,
             onFocus: () => {
                 FocusManager._currentFocusedComponent = storedComponent;
             }
         };
 
+        (component as any)._focusableComponentId = componentId;
         FocusManager._allFocusableComponents[componentId] = storedComponent;
 
-        if (!this._isRoot) {
-            this._myFocusableComponentIds[componentId] = true;
+        let withinRestrictionOwner: boolean = false;
+
+        for (let parent: FocusManager = this; parent; parent = parent._parent) {
+            parent._myFocusableComponentIds[componentId] = true;
+
+            if (FocusManager._currentRestrictionOwner === parent) {
+                withinRestrictionOwner = true;
+            }
+
+            if (parent._isFocusLimited) {
+                storedComponent.limitedCount++;
+            }
         }
 
-        if (FocusManager._currentRestrictionOwner && (this !== FocusManager._currentRestrictionOwner)) {
-            // New focusable element is mounted but it's not in the scope of the
-            // current view with restrictFocusWithin property.
-            FocusManager._restrictComponentFocus(componentId);
+        if (!withinRestrictionOwner && FocusManager._currentRestrictionOwner) {
+            storedComponent.restricted = true;
         }
 
-        (component as any)._focusableComponentId = componentId;
+        FocusManager._updateComponentFocusRestriction(storedComponent);
 
-        let el = ReactDOM.findDOMNode<HTMLElement>(component);
+        const el = ReactDOM.findDOMNode<HTMLElement>(component);
         if (el) {
             el.addEventListener('focus', storedComponent.onFocus);
         }
     }
 
     removeFocusableComponent(component: React.Component<any, any>) {
-        let componentId: string = (component as any)._focusableComponentId;
+        const componentId: string = (component as any)._focusableComponentId;
 
         if (componentId) {
-            let storedComponent: StoredFocusableComponent = FocusManager._allFocusableComponents[componentId];
+            const storedComponent: StoredFocusableComponent = FocusManager._allFocusableComponents[componentId];
 
-            let el = ReactDOM.findDOMNode<HTMLElement>(component);
+            const el = ReactDOM.findDOMNode<HTMLElement>(component);
             if (storedComponent && el) {
                 el.removeEventListener('focus', storedComponent.onFocus);
             }
 
             storedComponent.removed = true;
+            storedComponent.restricted = false;
+            storedComponent.limitedCount = 0;
 
-            FocusManager._removeComponentFocusRestriction(componentId);
-            delete this._myFocusableComponentIds[componentId];
+            FocusManager._updateComponentFocusRestriction(storedComponent);
+
+            delete storedComponent.callbacks;
+
+            for (let parent: FocusManager = this; parent; parent = parent._parent) {
+                delete parent._myFocusableComponentIds[componentId];
+            }
+
             delete FocusManager._allFocusableComponents[componentId];
             delete (component as any)._focusableComponentId;
         }
@@ -116,7 +145,7 @@ export class FocusManager {
         this._prevFocusedComponent = FocusManager._currentFocusedComponent;
 
         if (this._prevFocusedComponent && !this._prevFocusedComponent.removed) {
-            let el = ReactDOM.findDOMNode<HTMLElement>(this._prevFocusedComponent.component);
+            const el = ReactDOM.findDOMNode<HTMLElement>(this._prevFocusedComponent.component);
             if (el && el.blur) {
                 el.blur();
             }
@@ -124,50 +153,137 @@ export class FocusManager {
 
         Object.keys(FocusManager._allFocusableComponents).forEach(componentId => {
             if (!(componentId in this._myFocusableComponentIds)) {
-                FocusManager._restrictComponentFocus(componentId);
+                const storedComponent = FocusManager._allFocusableComponents[componentId];
+                storedComponent.restricted = true;
+                FocusManager._updateComponentFocusRestriction(storedComponent);
             }
         });
     }
 
-    release() {
-        // When the view is unmounted, restore the focus to the previous view with
-        // restrictFocusWithin or remove the restriction if there is no such view.
+    removeFocusRestriction() {
+        // Restore the focus to the previous view with restrictFocusWithin or
+        // remove the restriction if there is no such view.
         FocusManager._restrictionStack = FocusManager._restrictionStack.filter(focusManager => {
             return focusManager !== this;
         });
 
-        let prevFocusedComponent = this._prevFocusedComponent;
+        const prevFocusedComponent = this._prevFocusedComponent;
         delete this._prevFocusedComponent;
 
         if (FocusManager._currentRestrictionOwner === this) {
             FocusManager._removeFocusRestriction();
-            FocusManager._currentRestrictionOwner = FocusManager._restrictionStack[FocusManager._restrictionStack.length - 1];
+
+            const prevRestrictionOwner = FocusManager._restrictionStack.pop();
+            FocusManager._currentRestrictionOwner = undefined;
+
+            if (prevRestrictionOwner) {
+                prevRestrictionOwner.restrictFocusWithin();
+            }
 
             // If possible, focus the element which was focused before the restriction.
-            if (prevFocusedComponent && !prevFocusedComponent.removed && !prevFocusedComponent.restricted) {
-                let el = ReactDOM.findDOMNode<HTMLElement>(prevFocusedComponent.component);
+            if (prevFocusedComponent && !prevFocusedComponent.removed &&
+                    !prevFocusedComponent.restricted && !prevFocusedComponent.limitedCount) {
+                const el = ReactDOM.findDOMNode<HTMLElement>(prevFocusedComponent.component);
                 if (el && el.focus) {
                     el.focus();
                 }
             }
-
-            if (FocusManager._currentRestrictionOwner) {
-                FocusManager._currentRestrictionOwner.restrictFocusWithin();
-            }
         }
+    }
+
+    limitFocusWithin() {
+        if (this._isFocusLimited) {
+            return;
+        }
+
+        this._isFocusLimited = true;
+
+        Object.keys(this._myFocusableComponentIds).forEach(componentId => {
+            let storedComponent = FocusManager._allFocusableComponents[componentId];
+            storedComponent.limitedCount++;
+            FocusManager._updateComponentFocusRestriction(storedComponent);
+        });
+    }
+
+    removeFocusLimitation() {
+        if (!this._isFocusLimited) {
+            return;
+        }
+
+        Object.keys(this._myFocusableComponentIds).forEach(componentId => {
+            let storedComponent = FocusManager._allFocusableComponents[componentId];
+            storedComponent.limitedCount--;
+            FocusManager._updateComponentFocusRestriction(storedComponent);
+        });
+
+        this._isFocusLimited = false;
+    }
+
+    release() {
+        this.removeFocusRestriction();
+        this.removeFocusLimitation();
+    }
+
+    subscribe(component: React.Component<any, any>, callback: FocusableComponentStateCallback) {
+        const storedComponent = FocusManager._getStoredComponent(component);
+
+        if (storedComponent) {
+            if (!storedComponent.callbacks) {
+                storedComponent.callbacks = [];
+            }
+
+            storedComponent.callbacks.push(callback);
+        }
+    }
+
+    unsubscribe(component: React.Component<any, any>, callback: FocusableComponentStateCallback) {
+        const storedComponent = FocusManager._getStoredComponent(component);
+
+        if (storedComponent && storedComponent.callbacks) {
+            storedComponent.callbacks = storedComponent.callbacks.filter(cb => {
+                return cb !== callback;
+            });
+        }
+    }
+
+    isComponentFocusRestrictedOrLimited(component: React.Component<any, any>): boolean {
+        const storedComponent = FocusManager._getStoredComponent(component);
+        return storedComponent && (storedComponent.restricted || storedComponent.limitedCount > 0);
+    }
+
+    private static _getStoredComponent(component: React.Component<any, any>): StoredFocusableComponent {
+        const componentId: string = (component as any)._focusableComponentId;
+
+        if (componentId) {
+            return FocusManager._allFocusableComponents[componentId];
+        }
+
+        return null;
+    }
+
+    private static _callFocusableComponentStateChangeCallbacks(storedComponent: StoredFocusableComponent, restrictedOrLimited: boolean) {
+        if (!storedComponent.callbacks) {
+            return;
+        }
+
+        storedComponent.callbacks.forEach(callback => {
+            callback.call(storedComponent.component, restrictedOrLimited);
+        });
     }
 
     private static _removeFocusRestriction() {
         Object.keys(FocusManager._allFocusableComponents).forEach(componentId => {
-            FocusManager._removeComponentFocusRestriction(componentId);
+            let storedComponent = FocusManager._allFocusableComponents[componentId];
+            storedComponent.restricted = false;
+            FocusManager._updateComponentFocusRestriction(storedComponent);
         });
     }
 
     private static _setTabIndex(component: React.Component<any, any>, value: number): number {
-        let el = ReactDOM.findDOMNode<HTMLElement>(component);
+        const el = ReactDOM.findDOMNode<HTMLElement>(component);
 
         if (el) {
-            let prev = el.hasAttribute('tabindex') ? el.tabIndex : undefined;
+            const prev = el.hasAttribute('tabindex') ? el.tabIndex : undefined;
 
             if (value === undefined) {
                 if (prev !== undefined) {
@@ -183,22 +299,14 @@ export class FocusManager {
         return null;
     }
 
-    private static _restrictComponentFocus(componentId: string) {
-        let storedComponent = FocusManager._allFocusableComponents[componentId];
-
-        if (storedComponent && !storedComponent.restricted) {
+    private static _updateComponentFocusRestriction(storedComponent: StoredFocusableComponent) {
+        if ((storedComponent.restricted || (storedComponent.limitedCount > 0)) && !('origTabIndex' in storedComponent)) {
             storedComponent.origTabIndex = FocusManager._setTabIndex(storedComponent.component, -1);
-            storedComponent.restricted = true;
-        }
-    }
-
-    private static _removeComponentFocusRestriction(componentId: string) {
-        let storedComponent = FocusManager._allFocusableComponents[componentId];
-
-        if (storedComponent && storedComponent.restricted) {
+            FocusManager._callFocusableComponentStateChangeCallbacks(storedComponent, true);
+        } else if (!storedComponent.restricted && !storedComponent.limitedCount && ('origTabIndex' in storedComponent)) {
             FocusManager._setTabIndex(storedComponent.component, storedComponent.origTabIndex);
             delete storedComponent.origTabIndex;
-            storedComponent.restricted = false;
+            FocusManager._callFocusableComponentStateChangeCallbacks(storedComponent, false);
         }
     }
 }
@@ -240,7 +348,7 @@ export function applyFocusableComponentMixin(Component: any, isConditionallyFocu
         let origCallback = Component.prototype[methodName];
 
         Component.prototype[methodName] = function () {
-            let focusManager: FocusManager = this.context && this.context.focusManager;
+            let focusManager: FocusManager = this._focusManager || (this.context && this.context.focusManager);
 
             if (focusManager) {
                 action.call(this, focusManager, arguments);